Outsorce isPlayerOnAnyPlatform function
